+++ components/esp_insights/include/esp_insights.h
@@ -5,6 +5,7 @@
  */
 
 #pragma once
+
 #include <stdint.h>
 #include <esp_err.h>
 #include <esp_event_base.h>
@@ -200,6 +201,53 @@ void esp_insights_disable(void);
  * @return Pointer to a NULL terminated Node ID string.
  */
 const char *esp_insights_get_node_id(void);
+
+/**
+ * @brief Check if insights reporting is enabled
+ *
+ * @return true reporting is on
+ * @return false reporting is off
+ */
+bool esp_insights_is_reporting_enabled(void);
+
+/**
+ * @brief Turn on the Insights reporting
+ *
+ * @return esp_err_t ESP_OK on success, apt error otherwise
+ */
+esp_err_t esp_insights_reporting_enable();
+
+/**
+ * @brief Turn off the Insights repoting
+ *
+ * @return esp_err_t ESP_OK on success, apt error otherwise
+ * @note meta message if changed and the boot message will still be
+ *  sent as this information is critical for Insights working with the
+ *  cloud. You may disable insight completely using esp_insights_disable
+ */
+esp_err_t esp_insights_reporting_disable();
+
+/**
+ * @brief Encode and parse the command directly using esp-insight's parser
+ *
+ * This tests only if the parser is working as expected.
+ */
+esp_err_t esp_insights_test_cmd_handler();
+
+/**
+ * @brief Enable esp-insights command-response module
+ *
+ * This API registers esp-insights command parser which when data is received,
+ * parses it to filter out insights specific data, modifies configs accordingly,
+ * and prepares and gives response data to the module
+ *
+ * The \ref esp_insights_init takes care of initializing command response and
+ * enabling the same. In cases where, only esp_insights_enable is called, e.g.,
+ * ESP Rainmaker's app_insights module, user needs to call this API, before or
+ * after \ref esp_insights_enable
+ */
+esp_err_t esp_insights_cmd_resp_enable(void);
+
 #ifdef __cplusplus
 }
 #endif